requirements.txt
apturl==0.5.2
asgiref==3.4.1
asn1crypto==0.24.0
boto3==1.17.36
botocore==1.20.36
Brlapi==0.6.6
caffeine==2.9.4
certifi==2020.12.5
chardet==4.0.0
click==6.7
colorama==0.3.7
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==3.2.7
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.5
django-redis==5.0.0
django-session-timeout==0.1.0
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
ewmh==0.1.5
httplib2==0.9.2
idna==2.10
jmespath==0.10.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
menulibre==2.2.0
notify2==0.3
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==7.2.0
protobuf==3.0.0
psutil==5.4.2
psycopg2-binary==2.8.6
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
PyJWT==2.1.0
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.8.1
python-debian==0.1.32
python-decouple==3.4
python-http-client==3.3.2
python-xapp==1.0.0
python-xlib==0.20
pytz==2021.1
pyudev==0.21.0
pyxdg==0.25
PyYAML==3.12
redis==3.5.3
reportlab==3.4.0
requests==2.25.1
requests-unixsocket==0.1.5
s3transfer==0.3.6
SecretStorage==2.3.1
sendgrid==6.8.1
setproctitle==1.1.10
simplejson==3.13.2
six==1.15.0
sqlparse==0.4.2
starkbank-ecdsa==1.1.1
system-service==0.3
systemd-python==234
typing-extensions==3.10.0.2
ubuntu-drivers-common==0.0.0
unattended-upgrades==0.1
urllib3==1.26.3
wadllib==1.3.2
xkit==0.0.0
zope.interface==4.3.2
